refactor: update Providers import path to use ui directory
